Refactored PartsOfPlanTable.tsx - separated logic to a new file - usePartsOfPlanTableLogic.ts, in PartsOfPlanTable.tsx left only the view.
